Move filter group to new component
